requirements.txt
asgiref==3.7.2
Django==4.2.3
django-guardian==2.4.0
django-htmx==1.16.0
django-tenants==3.5.0
Pillow==10.0.0
psycopg2-binary==2.9.7
sqlparse==0.4.4
